Move `@react-native-community/art` to `peerDependencies`
`@react-native-community/art` is a project with native code and must be
linked into host application either manually or with autolinking.

Autolinking for a package only works as long as package is mentioned in
`package.json` of host application.

This looks like a case for `peerDependencies` since they are intended
for a situaions where some package has to be installed separately in
final `package.json` and not fetched automatically.